# Fix mvn surefire config: better test output