Mvn test output is now more detailed and helpful.

> Fix mvn surefire config: force surefire-junit47 as the surefire provider
> Previously surefire-junit3 was used.
> See also https://maven.apache.org/surefire/maven-surefire-plugin/examples/providers.html

Before:

```
-------------------------------------------------------
 T E S T S
-------------------------------------------------------
Running org.junit.tests.AllTests
Tests run: 940, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 8.58 sec - in org.junit.tests.AllTests

Results :

Tests run: 940, Failures: 0, Errors: 0, Skipped: 0
```

After:

```
[INFO] --- maven-surefire-plugin:2.19.1:test (default-test) @ junit ---

-------------------------------------------------------
 T E S T S
-------------------------------------------------------
Running org.junit.tests.assertion.AssertionTest
Tests run: 86, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.028 sec - in org.junit.tests.assertion.AssertionTest
Running org.junit.tests.assertion.ComparisonFailureTest
Tests run: 12, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.001 sec - in org.junit.tests.assertion.ComparisonFailureTest
Running org.junit.tests.assertion.MultipleFailureExceptionTest
Tests run: 4, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0 sec - in org.junit.tests.assertion.MultipleFailureExceptionTest
Running org.junit.tests.deprecated.JUnit4ClassRunnerTest
Tests run: 2, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0 sec - in org.junit.tests.deprecated.JUnit4ClassRunnerTest
Running org.junit.tests.description.AnnotatedDescriptionTest
Tests run: 6, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.01 sec - in org.junit.tests.description.AnnotatedDescriptionTest
Running org.junit.tests.description.SuiteDescriptionTest
Tests run: 2, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0 sec - in org.junit.tests.description.SuiteDescriptionTest
Running org.junit.tests.description.TestDescriptionMethodNameTest
Tests run: 11, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.002 sec - in org.junit.tests.description.TestDescriptionMethodNameTest
Running org.junit.tests.description.TestDescriptionTest
Tests run: 3, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.001 sec - in org.junit.tests.description.TestDescriptionTest
Running org.junit.experimental.categories.CategoriesAndParameterizedTest
Tests run: 3, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.011 sec - in org.junit.experimental.categories.CategoriesAndParameterizedTest
Running org.junit.experimental.categories.CategoryFilterFactoryTest
Tests run: 2, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.003 sec - in org.junit.experimental.categories.CategoryFilterFactoryTest
Running org.junit.experimental.categories.CategoryTest
Tests run: 23, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.055 sec - in org.junit.experimental.categories.CategoryTest
Running org.junit.experimental.categories.CategoryValidatorTest
Tests run: 5, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.005 sec - in org.junit.experimental.categories.CategoryValidatorTest
Running org.junit.experimental.categories.JavadocTest
Tests run: 2, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.007 sec - in org.junit.experimental.categories.JavadocTest
Running org.junit.experimental.categories.MultiCategoryTest
Tests run: 5, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.018 sec - in org.junit.experimental.categories.MultiCategoryTest
Running org.junit.tests.experimental.max.DescriptionTest
Tests run: 9, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.01 sec - in org.junit.tests.experimental.max.DescriptionTest
Running org.junit.tests.experimental.max.JUnit38SortingTest
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.017 sec - in org.junit.tests.experimental.max.JUnit38SortingTest
Running org.junit.tests.experimental.max.MaxStarterTest
Tests run: 16, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.535 sec - in org.junit.tests.experimental.max.MaxStarterTest
Running org.junit.tests.experimental.parallel.ParallelClassTest
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.003 sec - in org.junit.tests.experimental.parallel.ParallelClassTest
Running org.junit.tests.experimental.parallel.ParallelMethodTest
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.004 sec - in org.junit.tests.experimental.parallel.ParallelMethodTest
Running org.junit.tests.experimental.results.PrintableResultTest
Tests run: 2, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.034 sec - in org.junit.tests.experimental.results.PrintableResultTest
Running org.junit.tests.experimental.results.ResultMatchersTest
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0 sec - in org.junit.tests.experimental.results.ResultMatchersTest
Running org.junit.tests.experimental.theories.internal.AllMembersSupplierTest
Tests run: 10, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.001 sec - in org.junit.tests.experimental.theories.internal.AllMembersSupplierTest
Running org.junit.tests.experimental.theories.internal.ParameterizedAssertionErrorTest
Tests run: 7, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.01 sec - in org.junit.tests.experimental.theories.internal.ParameterizedAssertionErrorTest
Running org.junit.tests.experimental.theories.internal.SpecificDataPointsSupplierTest
Tests run: 6, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.006 sec - in org.junit.tests.experimental.theories.internal.SpecificDataPointsSupplierTest
Running org.junit.tests.experimental.theories.runner.FailingDataPointMethods
Tests run: 6, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.008 sec - in org.junit.tests.experimental.theories.runner.FailingDataPointMethods
Running org.junit.tests.experimental.theories.runner.SuccessfulWithDataPointFields$StaticPublicNonDataPoints
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0 sec - in org.junit.tests.experimental.theories.runner.SuccessfulWithDataPointFields$StaticPublicNonDataPoints
Running org.junit.tests.experimental.theories.runner.SuccessfulWithDataPointFields$OneTestTwoAnnotations
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0 sec - in org.junit.tests.experimental.theories.runner.SuccessfulWithDataPointFields$OneTestTwoAnnotations
Running org.junit.tests.experimental.theories.runner.SuccessfulWithDataPointFields$BeforeAndAfterEachTime
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0 sec - in org.junit.tests.experimental.theories.runner.SuccessfulWithDataPointFields$BeforeAndAfterEachTime
Running org.junit.tests.experimental.theories.runner.SuccessfulWithDataPointFields$DifferentTypesInConstructor
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0 sec - in org.junit.tests.experimental.theories.runner.SuccessfulWithDataPointFields$DifferentTypesInConstructor
Running org.junit.tests.experimental.theories.runner.SuccessfulWithDataPointFields$PositiveIntsWithMethodParams
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0 sec - in org.junit.tests.experimental.theories.runner.SuccessfulWithDataPointFields$PositiveIntsWithMethodParams
Running org.junit.tests.experimental.theories.runner.SuccessfulWithDataPointFields$PositiveIntsWithNegativeField
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0 sec - in org.junit.tests.experimental.theories.runner.SuccessfulWithDataPointFields$PositiveIntsWithNegativeField
Running org.junit.tests.experimental.theories.runner.SuccessfulWithDataPointFields$PositiveInts
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0 sec - in org.junit.tests.experimental.theories.runner.SuccessfulWithDataPointFields$PositiveInts
Running org.junit.tests.experimental.theories.runner.SuccessfulWithDataPointFields$NewObjectEachTime
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.002 sec - in org.junit.tests.experimental.theories.runner.SuccessfulWithDataPointFields$NewObjectEachTime
Running org.junit.tests.experimental.theories.runner.SuccessfulWithDataPointFields$BeforeAndAfterOnSameInstance
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0 sec - in org.junit.tests.experimental.theories.runner.SuccessfulWithDataPointFields$BeforeAndAfterOnSameInstance
Running org.junit.tests.experimental.theories.runner.SuccessfulWithDataPointFields$HasATwoParameterTheory
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0 sec - in org.junit.tests.experimental.theories.runner.SuccessfulWithDataPointFields$HasATwoParameterTheory
Running org.junit.tests.experimental.theories.runner.TheoriesPerformanceTest
Tests run: 1, Failures: 0, Errors: 0, Skipped: 1, Time elapsed: 0 sec - in org.junit.tests.experimental.theories.runner.TheoriesPerformanceTest
Running org.junit.tests.experimental.theories.runner.TypeMatchingBetweenMultiDataPointsMethod
Tests run: 2, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0 sec - in org.junit.tests.experimental.theories.runner.TypeMatchingBetweenMultiDataPointsMethod
Running org.junit.tests.experimental.theories.runner.UnsuccessfulWithDataPointFields
Tests run: 10, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.049 sec - in org.junit.tests.experimental.theories.runner.UnsuccessfulWithDataPointFields
Running org.junit.tests.experimental.theories.runner.WhenNoParametersMatch
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.009 sec - in org.junit.tests.experimental.theories.runner.WhenNoParametersMatch
Running org.junit.tests.experimental.theories.runner.WithAutoGeneratedDataPoints
Tests run: 4, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.163 sec - in org.junit.tests.experimental.theories.runner.WithAutoGeneratedDataPoints
Running org.junit.tests.experimental.theories.runner.WithDataPointMethod
Tests run: 4, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.008 sec - in org.junit.tests.experimental.theories.runner.WithDataPointMethod
Running org.junit.tests.experimental.theories.runner.WithExtendedParameterSources
Tests run: 8, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.039 sec - in org.junit.tests.experimental.theories.runner.WithExtendedParameterSources
Running org.junit.tests.experimental.theories.runner.WithNamedDataPoints
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0 sec - in org.junit.tests.experimental.theories.runner.WithNamedDataPoints
Running org.junit.tests.experimental.theories.runner.WithOnlyTestAnnotations
Tests run: 4, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.03 sec - in org.junit.tests.experimental.theories.runner.WithOnlyTestAnnotations
Running org.junit.tests.experimental.theories.runner.WithParameterSupplier
Tests run: 4, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0 sec - in org.junit.tests.experimental.theories.runner.WithParameterSupplier
Running org.junit.tests.experimental.theories.runner.WithUnresolvedGenericTypeVariablesOnTheoryParms
Tests run: 9, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.07 sec - in org.junit.tests.experimental.theories.runner.WithUnresolvedGenericTypeVariablesOnTheoryParms
Running org.junit.tests.experimental.theories.ParameterSignatureTest
Tests run: 5, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.004 sec - in org.junit.tests.experimental.theories.ParameterSignatureTest
Running org.junit.tests.experimental.theories.TestedOnSupplierTest
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.001 sec - in org.junit.tests.experimental.theories.TestedOnSupplierTest
Running org.junit.tests.experimental.theories.AssumingInTheoriesTest
Tests run: 2, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.002 sec - in org.junit.tests.experimental.theories.AssumingInTheoriesTest
Running org.junit.tests.experimental.theories.PotentialAssignmentTest
Tests run: 4, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.002 sec - in org.junit.tests.experimental.theories.PotentialAssignmentTest
Running org.junit.tests.experimental.AssumptionTest
Tests run: 18, Failures: 0, Errors: 0, Skipped: 4, Time elapsed: 0.017 sec - in org.junit.tests.experimental.AssumptionTest
Running org.junit.tests.experimental.MatcherTest
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.021 sec - in org.junit.tests.experimental.MatcherTest
Running org.junit.tests.experimental.theories.extendingwithstubs.StubbedTheoriesTest
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.004 sec - in org.junit.tests.experimental.theories.extendingwithstubs.StubbedTheoriesTest
Running org.junit.internal.builders.AnnotatedBuilderTest
Tests run: 6, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.003 sec - in org.junit.internal.builders.AnnotatedBuilderTest
Running org.junit.internal.runners.ErrorReportingRunnerTest
Tests run: 4, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.002 sec - in org.junit.internal.runners.ErrorReportingRunnerTest
Running org.junit.internal.runners.statements.FailOnTimeoutTest
Tests run: 8, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.79 sec - in org.junit.internal.runners.statements.FailOnTimeoutTest
Running org.junit.internal.MethodSorterTest
Tests run: 6, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0 sec - in org.junit.internal.MethodSorterTest
Running org.junit.internal.matchers.StacktracePrintingMatcherTest
Tests run: 3, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0 sec - in org.junit.internal.matchers.StacktracePrintingMatcherTest
Running org.junit.internal.matchers.ThrowableCauseMatcherTest
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0 sec - in org.junit.internal.matchers.ThrowableCauseMatcherTest
Running org.junit.tests.junit3compatibility.AllTestsTest
Tests run: 5, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.014 sec - in org.junit.tests.junit3compatibility.AllTestsTest
Running org.junit.tests.junit3compatibility.ClassRequestTest
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0 sec - in org.junit.tests.junit3compatibility.ClassRequestTest
Running org.junit.tests.junit3compatibility.ForwardCompatibilityPrintingTest
Tests run: 2, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.019 sec - in org.junit.tests.junit3compatibility.ForwardCompatibilityPrintingTest
Running org.junit.tests.junit3compatibility.ForwardCompatibilityTest
Tests run: 12, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.01 sec - in org.junit.tests.junit3compatibility.ForwardCompatibilityTest
Running org.junit.tests.junit3compatibility.InitializationErrorForwardCompatibilityTest
Tests run: 3, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.002 sec - in org.junit.tests.junit3compatibility.InitializationErrorForwardCompatibilityTest
Running org.junit.tests.junit3compatibility.JUnit38ClassRunnerTest
Tests run: 7, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.034 sec - in org.junit.tests.junit3compatibility.JUnit38ClassRunnerTest
Running org.junit.tests.junit3compatibility.OldTestClassAdaptingListenerTest
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0 sec - in org.junit.tests.junit3compatibility.OldTestClassAdaptingListenerTest
Running junit.tests.framework.TestCaseTest
Tests run: 14, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.005 sec - in junit.tests.framework.TestCaseTest
Running junit.tests.framework.SuiteTest
Tests run: 11, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.041 sec - in junit.tests.framework.SuiteTest
Running junit.tests.framework.TestListenerTest
Tests run: 3, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.001 sec - in junit.tests.framework.TestListenerTest
Running junit.tests.framework.AssertionFailedErrorTest
Tests run: 3, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.001 sec - in junit.tests.framework.AssertionFailedErrorTest
Running junit.tests.framework.AssertTest
Tests run: 16, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.004 sec - in junit.tests.framework.AssertTest
Running junit.tests.framework.TestImplementorTest
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.001 sec - in junit.tests.framework.TestImplementorTest
Running junit.tests.framework.NoArgTestCaseTest
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0 sec - in junit.tests.framework.NoArgTestCaseTest
Running junit.tests.framework.ComparisonCompactorTest
Tests run: 19, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.004 sec - in junit.tests.framework.ComparisonCompactorTest
Running junit.tests.framework.ComparisonFailureTest
Tests run: 4, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.001 sec - in junit.tests.framework.ComparisonFailureTest
Running junit.tests.framework.DoublePrecisionAssertTest
Tests run: 7, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.001 sec - in junit.tests.framework.DoublePrecisionAssertTest
Running junit.tests.framework.FloatAssertTest
Tests run: 8, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.004 sec - in junit.tests.framework.FloatAssertTest
Running junit.tests.runner.StackFilterTest
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.001 sec - in junit.tests.runner.StackFilterTest
Running junit.tests.runner.ResultTest
Tests run: 2, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.107 sec - in junit.tests.runner.ResultTest
Running junit.tests.runner.BaseTestRunnerTest
Tests run: 2, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.002 sec - in junit.tests.runner.BaseTestRunnerTest
Running junit.tests.runner.TextFeedbackTest
Tests run: 5, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.004 sec - in junit.tests.runner.TextFeedbackTest
Running junit.tests.runner.TextRunnerSingleMethodTest
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.001 sec - in junit.tests.runner.TextRunnerSingleMethodTest
Running junit.tests.runner.TextRunnerTest
Tests run: 4, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.31 sec - in junit.tests.runner.TextRunnerTest
Running junit.tests.extensions.ExtensionTest
Tests run: 4, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.001 sec - in junit.tests.extensions.ExtensionTest
Running junit.tests.extensions.ActiveTestTest
Tests run: 4, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.234 sec - in junit.tests.extensions.ActiveTestTest
Running junit.tests.extensions.RepeatedTestTest
Tests run: 4, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0 sec - in junit.tests.extensions.RepeatedTestTest
Running org.junit.tests.junit3compatibility.SuiteMethodTest
Tests run: 5, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0 sec - in org.junit.tests.junit3compatibility.SuiteMethodTest
Running org.junit.tests.listening.ListenerTest
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0 sec - in org.junit.tests.listening.ListenerTest
Running org.junit.tests.listening.RunnerTest
Tests run: 3, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0 sec - in org.junit.tests.listening.RunnerTest
Running org.junit.tests.listening.TestListenerTest
Tests run: 3, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0 sec - in org.junit.tests.listening.TestListenerTest
Running org.junit.tests.listening.TextListenerTest
Tests run: 3, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.001 sec - in org.junit.tests.listening.TextListenerTest
Running org.junit.tests.listening.UserStopTest
Tests run: 2, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.001 sec - in org.junit.tests.listening.UserStopTest
Running org.junit.tests.manipulation.FilterableTest
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.002 sec - in org.junit.tests.manipulation.FilterableTest
Running org.junit.tests.manipulation.FilterTest
…