@alb-i986 2016-05-12
> Do you know why the test count is different now?

It seems to be due to the change [surefire-junit3](https://github.com/apache/maven-surefire/tree/master/surefire-providers/surefire-junit3) -> [surefire-junit47](https://github.com/apache/maven-surefire/tree/master/surefire-providers/surefire-junit47).

@Tibor17 may have some clue