@alb-i986 2016-05-12
BTW if I run org.junit.tests.AllTests with IDEA, by clicking on the class name and selecting Run test (thus bypassing mvn), I get the same result as the one I get with surefire-junit47. That's good news, I think.

Also, about the skipped tests, I've found a bug:
when using `@Test(expected = AssumptionViolatedException.class)` the test will actually be marked as skipped, as if the AssumptionViolatedException were thrown up in the stack.
But maybe that's expected. I've temporarily fixed this by using a try/catch block. I'll open an issue.

With this we're now at 

    Tests run: 942, Failures: 0, Errors: 0, Skipped: 4

Then I see there is something wrong with this test:

```
    @Test(expected = IllegalArgumentException.class)
    public void assumeWithExpectedException() {
        assumeTrue(false);
    }
```

(in AssumptionTest)

It's being marked as skipped. Again, I'll open a separate issue.