@kcooney 2016-05-13
I have a slight preference for 6413e99 as a separate pull. Is it too much trouble?