@alb-i986 2016-05-14
> I have a slight preference for 6413e99 as a separate pull. Is it too much trouble?

 Of course not. Will do that now.