@marcphilipp 2016-05-14
I really like this and would like to merge it.

@kcooney Any objections?